reafactor: extract middleware
